+++ Tortuga.Chain/Tortuga.Chain.SqlServer.MDS/SqlServer/AbstractSqlServerMetadataCache.cs
@@ -51,6 +51,8 @@ partial class OleDbSqlServerMetadataCache : AbstractOleDbSqlServerMetadataCache
 
 	internal readonly ConcurrentDictionary<SqlServerObjectName, UserDefinedTableTypeMetadata<SqlServerObjectName, AbstractDbType>> m_UserDefinedTableTypes = new();
 
+	readonly ConcurrentDictionary<int, SqlServerObjectName> m_ObjectIdTableMap = new();
+
 	internal string? m_DatabaseName;
 	internal string? m_DefaultSchema;
 
@@ -183,6 +185,7 @@ partial class OleDbSqlServerMetadataCache : AbstractOleDbSqlServerMetadataCache
 		m_UserDefinedTableTypes.Clear();
 		m_ScalarFunctions.Clear();
 		m_ScalarFunctions.Clear();
+		m_ObjectIdTableMap.Clear();
 		m_DefaultSchema = null;
 	}
 
@@ -298,4 +301,24 @@ partial class OleDbSqlServerMetadataCache : AbstractOleDbSqlServerMetadataCache
 	/// Get the maximum number of rows in a single SQL statement's Values clause.
 	/// </summary>
 	public override int? MaxRowsPerValuesClause => 1000;
+
+
+	/// <summary>
+	/// Gets the name of the table or view.
+	/// </summary>
+	/// <param name="objectId">The object identifier.</param>
+	public SqlServerObjectName GetTableOrViewName(int objectId)
+	{
+		return m_ObjectIdTableMap.GetOrAdd(objectId, x => GetTableOrViewNameInternal(x));
+	}
+
+	/// <summary>
+	/// Gets the table or view from its object_id.
+	/// </summary>
+	/// <param name="objectId">The object identifier.</param>
+	public SqlServerTableOrViewMetadata<AbstractDbType> GetTableOrView(int objectId)
+	{
+		return GetTableOrView(GetTableOrViewName(objectId));
+	}
+
 }